Additional qseq validation and ambiguous nucleotide conversion (#2)
* Additional qseq validation and ambiguous nucleotide conversion

added additional validation parameters to verify that the file being read in is a valid qseq file. Specifically, qseq_to_fastq now confirms that the sequence and quality score are equal in length and confirms that the x and y coordinates from the flow cell are numerals. Additionally, ambiguous nucleotides now appear as 'N' in fastq rather than '.', as in qseq

* removed test files

removed test files